backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
SQLAlchemy==2.0.35
pydantic==1.10.17
python-multipart==0.0.9
grpcio==1.66.1
grpcio-tools==1.66.1
